LP64
git-svn-id: https://asn1c.svn.sourceforge.net/svnroot/asn1c/trunk@560 59561ff5-6e30-0410-9f3c-9617f08c8826

+++ skeletons/NumericString.c
@@ -54,10 +54,10 @@ NumericString_constraint(asn_TYPE_descriptor_t *td, const void *sptr,
 				continue;
 			}
 			_ASN_ERRLOG(app_errlog, app_key,
-				"%s: value byte %d (%d) "
+				"%s: value byte %ld (%d) "
 				"not in NumericString alphabet (%s:%d)",
 				td->name,
-				(buf - st->buf) + 1,
+				(long)((buf - st->buf) + 1),
 				*buf,
 				__FILE__, __LINE__);
 			return -1;
